requirements.txt
sqlmodel==0.0.16
spotipy==2.23.0
pokebase==1.3.0
python-dotenv==1.0.1

